# src/controllers/matching/matching.controller.ts
/**
 * Matching Controller
 * Exposes the optimized matching engine via REST API
 *
 * Uses Clarke-Wright + ALNS algorithms for optimal order-to-driver matching
 * with multi-round rejection handling
 */

import {
  Controller,
  Get,
  Post,
  Route,
  Response,
  Tags,
  Query,
  Body,
} from "tsoa";
import { matchOrders } from "../../services/matching/matchingEngine";
import { getDistance } from "../../services/routing/mapboxClient";
import { orderAssignmentService } from "../../services/assignment/order-assignment.service";

interface RouteInfo {
  driverId: string;
  driverName: string;
  orderCount: number;
  totalDistance: number;
  distancePerOrder: number;
}

interface MatchingResponse {
  success: boolean;
  message: string;
  routes: RouteInfo[];
  summary: {
    totalRoutes: number;
    totalOrders: number;
    totalAssigned: number;
    totalDistance: number;
    computationTimeMs: number;
    timestamp: string;
  };
}

@Route("matching")
@Tags("Matching Engine")
export class MatchingController extends Controller {
  /**
   * Execute the divide-and-conquer matching engine
   *
   * Triggers the three-stage matching algorithm to:
   * - Assign all pending orders to available drivers
   * - Optimize delivery routes for each driver
   * - Return detailed route information and metrics
   *
   * @param verbose Include detailed route waypoints in response (default: false)
   * @returns MatchingResponse with optimized routes and metrics
   */
  @Post("optimize")
  @Response<MatchingResponse>(200, "Matching completed successfully")
  @Response<{ error: string }>(400, "No pending orders or available drivers")
  @Response<{ error: string }>(500, "Matching engine error")
  async optimizeMatching(
    @Query() verbose: boolean = false
  ): Promise<MatchingResponse> {
    console.log("📍 API: Optimize matching request received");
    const startTime = Date.now();

    const routes = await matchOrders(false);

    const totalDistance = routes.reduce((sum, r) => sum + r.totalDistance, 0);
    const totalOrders = routes.reduce(
      (sum, r) => sum + r.metrics.orderCount,
      0
    );
    const computationTimeMs = Date.now() - startTime;

    const responseMessage = `Matching completed: ${totalOrders} orders assigned across ${routes.length} routes${
      verbose ? " with detailed waypoints" : ""
    }`;

    return {
      success: true,
      message: responseMessage,
      routes: routes.map((r) => ({
        driverId: r.driverId,
        driverName: r.driverName,
        orderCount: r.metrics.orderCount,
        totalDistance: r.totalDistance,
        distancePerOrder: r.metrics.distancePerOrder,
      })),
      summary: {
        totalRoutes: routes.length,
        totalOrders,
        totalAssigned: totalOrders,
        totalDistance,
        computationTimeMs,
        timestamp: new Date().toISOString(),
      },
    };
  }

  /**
   * Health check for matching service
   */
  @Get("health")
  @Response<{ status: string }>(200, "Service is healthy")
  async healthCheck(): Promise<{ status: string }> {
    return { status: "Matching service is healthy" };
  }

  @Post("route")
  async getRoute(
    @Body() body: { coordinates: number[][]; profile?: string }
  ): Promise<any> {
    if (!body.coordinates || body.coordinates.length < 2) {
      throw new Error("At least 2 coordinates are required");
    }

    const from = { lat: body.coordinates[0][1], lng: body.coordinates[0][0] };
    const to = { lat: body.coordinates[1][1], lng: body.coordinates[1][0] };

    // Cast profile to any to bypass strict literal type check, or import MapboxProfile
    return await getDistance(from, to, body.profile as any);
  }

  /**
   * Accept all OFFERED assignments (testing utility)
   */
  @Post("accept-all")
  @Response<{ success: boolean; acceptedCount: number; message: string }>(
    200,
    "All assignments accepted"
  )
  async acceptAll(): Promise<{
    success: boolean;
    acceptedCount: number;
    message: string;
  }> {
    const acceptedCount = await orderAssignmentService.acceptAllAssignments();
    return {
      success: true,
      acceptedCount,
      message: `Accepted ${acceptedCount} assignments`,
    };
  }

  /**
   * Reject all OFFERED assignments (testing utility)
   */
  @Post("reject-all")
  @Response<{ success: boolean; rejectedCount: number; message: string }>(
    200,
    "All assignments rejected"
  )
  async rejectAll(
    @Body() body?: { reason?: string }
  ): Promise<{ success: boolean; rejectedCount: number; message: string }> {
    const rejectedCount = await orderAssignmentService.rejectAllAssignments(
      body?.reason || "Bulk rejection for testing"
    );
    return {
      success: true,
      rejectedCount,
      message: `Rejected ${rejectedCount} assignments`,
    };
  }
}