# src/generated/swagger.json
{
	"components": {
...
		"schemas": {
...
			"RouteInfo": {
				"properties": {
					"driverId": {
						"type": "string"
					},
					"driverName": {
						"type": "string"
					},
					"orderCount": {
						"type": "number",
						"format": "double"
					},
					"totalDistance": {
						"type": "number",
						"format": "double"
					},
					"distancePerOrder": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"driverId",
					"driverName",
					"orderCount",
					"totalDistance",
					"distancePerOrder"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MatchingResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"routes": {
						"items": {
							"$ref": "#/components/schemas/RouteInfo"
						},
						"type": "array"
					},
					"summary": {
						"properties": {
							"timestamp": {
								"type": "string"
							},
							"computationTimeMs": {
								"type": "number",
								"format": "double"
							},
							"totalDistance": {
								"type": "number",
								"format": "double"
							},
							"totalAssigned": {
								"type": "number",
								"format": "double"
							},
							"totalOrders": {
								"type": "number",
								"format": "double"
							},
							"totalRoutes": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"timestamp",
							"computationTimeMs",
							"totalDistance",
							"totalAssigned",
							"totalOrders",
							"totalRoutes"
						],
						"type": "object"
					}
				},
				"required": [
					"success",
					"message",
					"routes",
					"summary"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "EdgeRun API",
		"version": "1.0.0",
		"description": "EdgeRun Last-Mile Delivery API - MVP for food delivery driver-order matching",
		"license": {
			"name": {
				"name": "MIT"
			}
		},
		"contact": {
			"name": "EdgeRun Support"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/matching/optimize": {
			"post": {
				"operationId": "OptimizeMatching",
				"responses": {
					"200": {
						"description": "MatchingResponse with optimized routes and metrics",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MatchingResponse"
								}
							}
						}
					},
					"400": {
						"description": "No pending orders or available drivers",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Matching engine error",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Execute the divide-and-conquer matching engine\n\nTriggers the three-stage matching algorithm to:\n- Assign all pending orders to available drivers\n- Optimize delivery routes for each driver\n- Return detailed route information and metrics",
				"tags": [
					"Matching Engine"
				],
				"security": [],
				"parameters": [
					{
						"description": "Include detailed route waypoints in response (default: false)",
						"in": "query",
						"name": "verbose",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					}
				]
			}
		},
...
		"/matching/route": {
			"post": {
				"operationId": "GetRoute",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Matching Engine"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"profile": {
										"type": "string"
									},
									"coordinates": {
										"items": {
											"items": {
												"type": "number",
												"format": "double"
											},
											"type": "array"
										},
										"type": "array"
									}
								},
								"required": [
									"coordinates"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
		"/matching/reject-all": {
			"post": {
				"operationId": "RejectAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"rejectedCount": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"rejectedCount",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Reject all OFFERED assignments (testing utility)",
				"tags": [
					"Matching Engine"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"reason": {
										"type": "string"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}